Move `index` to `fsm`
