+++ outlines/fsm/types.py
